pyproject.toml
[project]
name = "qualia"
version = "0.1.0"
description = "My qualia (and/or assistant) powerd by GPTs"
authors = [
    { name = "Yusei Ito", email = "me@yuseiito.com" }
]
dependencies = [
    "discord-py>=2.4.0",
    "openai>=1.61.1",
    "boto3>=1.36.14",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/qualia"]
